simple CHIP-8 syntax and indentation rules
this is my first attempt at writing syntax and indentation rules for
vim, so things are pretty sloppy. my goal was to cover the basics of
CHIP-8 from the manual and tutorials that I found at
https://github.com/JohnEarnest/Octo

I have left certain function keywords like sprite and jump out of the
default highlighting because it looked a bit too cluttered, but if these
are desired the highlight link rules can be uncommented.

there are also a number of SuperCHIP keywords but I haven't tested these