add syntax highlighting for the 'long' keyword
the original release chose not to include a lot of keywords (such as
save and load) by default, for both chip8 and superChip8. For now I'll
do the same for the XO-Chip but certain keywords like 'long' should
be included by default. since long addresses can be up to 4 hex digits
matched numbers need to support this length as well.