highlight 'audio' action
becuase the audio command is a stand alone action and not a function
with arguments it stands to differentiate it, just like 'exit' and
'return'. unfortunately, this means that the enable it means
uncommenting two lines of code instead of one. maybe I'll wrap them up
in a single conditional to include all the needed things at once in the
future.