distinguish between SuperChip and XO-Chip syntax more clearly
in order to specifically include SuperChip or XO-Chip instructions I had
the highlight rules commented out and instructed people to just
uncomment these liens. this was fine when it was just one line it was
starting to get more complex. now all of the syntax rules and linking
happens under a conditional and only one clear line has to be changed
for inclusion again.