indent correctly on nested blocks
when there are nested loops or nested if...begin statements the
autoindentation fell apart. I revamped the prevMatch function to handle
a start and end condition for the block that it searches for. I also
tweaked it a bit to be more vimscripty.

at least I *think* it's more vimscripty.